Fix warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ cores/arduino/pins_arduino.h
@@ -35,17 +35,17 @@ extern "C" {
 #define NOT_AN_INTERRUPT            NC // -1
 #define DEND                        PEND
 #define NUM_DIGITAL_PINS            ((uint32_t)DEND)
-#define NUM_ANALOG_INPUTS           (AEND-A0)
+#define NUM_ANALOG_INPUTS           ((uint32_t)(AEND-A0))
 
 // Convert a digital pin number Dxx to a PinName PX_n
 // Note: Analog pin is also a digital pin.
-#define digitalPinToPinName(p)      ((p < NUM_DIGITAL_PINS) ? digitalPin[p] : NC)
+#define digitalPinToPinName(p)      (((uint32_t)p < NUM_DIGITAL_PINS) ? digitalPin[p] : NC)
 // Convert a PinName PX_n to a digital pin number
 uint32_t pinNametoDigitalPin(PinName p);
 
 // Convert an analog pin number to a digital pin number
 // Used by analogRead api to have A0 == 0
-#define analogInputToDigitalPin(p)  ((p < NUM_ANALOG_INPUTS) ? (p+A0) : p)
+#define analogInputToDigitalPin(p)  (((uint32_t)p < NUM_ANALOG_INPUTS) ? (p+A0) : p)
 // Convert an analog pin number Axx to a PinName PX_n
 #define analogInputToPinName(p)     (digitalPinToPinName(analogInputToDigitalPin(p)))
 // All pins could manage EXTI
